started working on straight

+++ src/calculates.h
@@ -1,9 +1,9 @@
 class calculate {
     public:
-    int rotationTime(int diam, int distance, int rot_speed);
+    int rotationTime(int diam, int distance, float rot_speed);
 };
 
-int calculate::rotationTime (int diam, int distance, int rot_speed){
+int calculate::rotationTime (int diam, int distance, float rot_speed){
     float rot_numb = distance / (diam * 3.14);
     int time = (rot_numb / rot_speed) * 1000;
     return time;
